+++ CHANGELOG.md
@@ -13,3 +13,12 @@
 # 0.1.3
 
 - Force white color for contents to prevent color changing by custom ThemeData
+
+# 0.2.0
+
+- **Breaking Change** Change title and description to `Text`
+- **Breaking Change** Change `type` argument to `color`- make title optional
+- add color to icon
+- add color to action
+- add `CustomSheetColor` for Colors customisation
+